requirements.txt
streamlit
langchain
langchain-openai
langgraph
openai
python-dotenv
Pillow
requests